webapp/requirements.txt
Django==1.11.6
mysqlclient==1.3.12
pytz==2017.2
djangorestframework==3.7.3
markdown==2.6.9
django-filter==1.1.0
pyyaml==3.12
django-rest-auth==0.9.2
simplejson==3.12.0